requirements.txt
beautifulsoup4==4.10.0
finnhub_python==2.4.7
influxdb_client==1.25.0
requests==2.26.0
websocket_client==1.2.3
